Moved activities to activity folder and created navigation from main activity to reports list activity
