+++ app/src/main/AndroidManifest.xml
@@ -11,16 +11,15 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup,GoogleAppIndexingWarning">
-        <activity android:name=".ChatActivity">
+        <activity android:name=".activities.ChatActivity"/>
+        <activity android:name=".activities.LoginActivity" />
+        <activity android:name=".activities.ReportsListActivity" />
+        <activity android:name=".activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".ReportsListActivity" />
-        <activity android:name=".MainActivity"></activity>
+            </intent-filter></activity>
     </application>
 
 </manifest>
