+++ app/src/main/java/com/mobile/andrada/reportstuff/activities/ChatActivity.java
@@ -1,13 +1,15 @@
-package com.mobile.andrada.reportstuff;
+package com.mobile.andrada.reportstuff.activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class LoginActivity extends AppCompatActivity {
+import com.mobile.andrada.reportstuff.R;
+
+public class ChatActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
+        setContentView(R.layout.activity_chat);
     }
 }
