+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".activities.MainActivity">
 
     <LinearLayout
         android:layout_width="203dp"
@@ -22,17 +22,17 @@
         app:layout_constraintTop_toTopOf="parent">
 
         <Button
-            android:id="@+id/openedReportsButton"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Opened Reports" />
-
-        <Button
             android:id="@+id/newReportsButton"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:baselineAligned="false"
             android:text="New Reports" />
+
+        <Button
+            android:id="@+id/openedReportsButton"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Opened Reports" />
 
         <Button
             android:id="@+id/closedReportsButton"
